requirements.txt
google-cloud-bigtable==0.22.0
google-cloud-core==0.22.1
flask